MAILET-115 Extract replaceInternetAddresses from AbstractRedirect to SpecialAddressesUtils
